ajustes backed

+++ app/Services/SupplierService.php
@@ -6,6 +6,7 @@ use App\DTOs\CreateSupplierDTO;
 use App\DTOs\SupplierResponseDTO;
 use App\Models\Supplier;
 use App\Repositories\Contracts\SupplierRepositoryInterface;
+use Illuminate\Http\Exceptions\HttpResponseException;
 use Illuminate\Pagination\LengthAwarePaginator;
 
 class SupplierService
@@ -27,6 +28,10 @@ class SupplierService
     }
     public function createSupplier(CreateSupplierDTO $dto): Supplier
     {
+        $existingSupplier = $this->supplierRepository->getSupplierByCnpj($dto->supplier_cnpj);
+        if ($existingSupplier) {
+            throw new HttpResponseException(response()->json(['message' => 'Fornecedor com CNPJ '.$dto->supplier_cnpj.' já cadastrado'], 422));
+        }
         return $this->supplierRepository->createSupplier($dto);
     }
     public function updateSupplier(int $id, CreateSupplierDTO $dto): Supplier
@@ -35,6 +40,13 @@ class SupplierService
     }
     public function destroySupplier(int $id): void
     {
+        $supplier = $this->supplierRepository->showSupplier($id);
+        if ($supplier->fuelSuppliers) {
+            throw new HttpResponseException(response()->json(['message' => 'Fornecedor possui abastecimentos cadastrados'], 422));
+        }
+        if ($supplier->maintenances) {
+            throw new HttpResponseException(response()->json(['message' => 'Fornecedor possui manutenções cadastradas'], 422));
+        }
         $this->supplierRepository->destroySupplier($id);
     }
     public function showSupplier(int $id): Supplier
